Remove evaluate-related functions to a separate file
Evaluate has a handful of different functions that provide the
necessary logic. They were clogging up the `orbs.ts` file however.
All evaluate-related functions (including evaluate) were moved to
a `src/evaluate.ts` file. I like the call syntax for
`orbs.evaluate()` though, so evaluate is imported and exported in
the `src/orbs.ts` file.
